✨feat: Pin 버튼 정의 위치 이동
