+++ src/app/log/record/components/LogRecordRecording/LogRecordRecording.view.tsx
@@ -1,17 +1,14 @@
 import { Button } from "@/components";
 import * as S from "./LogRecordRecording.styles";
 import Theme, { FONT_SIZE, FONT_WEIGHT } from "@/styles/theme";
-import { Pin } from "@/components/icons";
 import { MasilRecordRequest } from "@/types/Request";
 
 interface LogRecordRecordingViewProps {
   logData: MasilRecordRequest;
-  handleClickCreatePin: () => void;
   handleClickCompleteRecord: () => void;
 }
 
 const LogRecordRecordingView = ({
-  handleClickCreatePin,
   handleClickCompleteRecord,
   logData,
 }: LogRecordRecordingViewProps) => {
@@ -25,14 +22,6 @@ const LogRecordRecordingView = ({
 
   return (
     <S.LogRecordActionLayout>
-      <Button
-        style={{ marginBottom: "1.6rem" }}
-        variant="neumorp"
-        onClickHandler={handleClickCreatePin}
-      >
-        <Pin />
-      </Button>
-
       <S.LogRecordActionContainer>
         <S.LogRecordInfoContainer>
           <S.LogRecordInfo>{`${hour} : ${min} : ${sec}`}</S.LogRecordInfo>
@@ -43,7 +32,7 @@ const LogRecordRecordingView = ({
           buttonColor={Theme.lightTheme.green_500}
           variant="neumorp"
           textColor={Theme.lightTheme.white}
-          style={{ fontWeight: FONT_WEIGHT.BOLD, opacity: 0.9, fontSize: FONT_SIZE.LARGE }}
+          style={{ fontWeight: FONT_WEIGHT.BOLD, fontSize: FONT_SIZE.LARGE }}
           width={"100%"}
           onClickHandler={handleClickCompleteRecord}
         >
